% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{col_vis_modal}
\alias{col_vis_modal}
\title{Table column visibility selection modal dialog}
\usage{
col_vis_modal(failed = FALSE, ns, dictionary)
}
\arguments{
\item{failed}{Logical}

\item{ns}{session ns}

\item{dictionary}{list}
}
\description{
Table column visibility selection modal dialog
}
