% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_scenario}
\alias{get_scenario}
\title{Get the current data scenario}
\usage{
get_scenario(viz_shape, agg_level)
}
\arguments{
\item{viz_shape}{The shape to visualize, plots or polygons}

\item{agg_level}{The breakdown level, plot, species...}
}
\description{
Return the current scenario as character
}
