% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{infopanel_climaplot_gen}
\alias{infopanel_climaplot_gen}
\title{generate the plot for the info panel, in climatic mode}
\usage{
infopanel_climaplot_gen(data, color, click)
}
\arguments{
\item{data}{list with clima data, filtered and no filtered (for plots)}

\item{color}{input}

\item{click}{input}
}
\description{
generate the plot for the info panel, in climatic mode
}
