% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{label_getter}
\alias{label_getter}
\title{label getter}
\usage{
label_getter(ifndb, lang = "esp", label_id, label_sub1 = NULL,
  label_sub2 = NULL)
}
\arguments{
\item{ifndb}{pool}

\item{lang}{character}

\item{label_id}{character}

\item{label_sub1}{character}

\item{label_sub2}{character}
}
\description{
label getter
}
