% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{map_modificator}
\alias{map_modificator}
\title{modifcate the map based on vis inputs}
\usage{
map_modificator(data_scenario, scenario, ifn, inverse_pal, color, mida,
  tipo_grup_func, grup_func, statistic, admin_div, agg_level,
  updateProgress = NULL)
}
\arguments{
\item{data_scenario}{tbl_sql object}

\item{scenario}{character indicating the scenario}

\item{ifn}{Input}

\item{inverse_pal}{Input}

\item{color}{Input}

\item{mida}{Input}

\item{tipo_grup_func}{Input}

\item{grup_func}{Input}

\item{statistic}{Input}

\item{updateProgress}{progress function to monitorize the map drawing}
}
\description{
this returns the leaflet proxy object to modify the map
}
