% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{mod_dataInput}
\alias{mod_dataInput}
\alias{mod_data}
\title{mod_dataInput and mod_data}
\usage{
mod_dataInput(id)

mod_data(input, output, session, mod_buttons)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{mod_buttons}{reactives from the mod_buttons module}
}
\description{
A shiny module to create and populate the data inputs

mod_data server function
}
