% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infopanelUI.R
\name{mod_infopanel}
\alias{mod_infopanel}
\title{mod_infopanel server function}
\usage{
mod_infopanel(input, output, session, mod_data, mod_map, mod_advancedFilters,
  ifndb)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{mod_data}{reactive with the reactive data and the data inputs}

\item{mod_map}{reactive with the map events from the map module}

\item{mod_advancedFilters}{reactive with the reactive values from
advancedFilters module}

\item{ifndb}{pool with database connection object}
}
\description{
mod_infopanel server function
}
