% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infopanelUI.R
\name{mod_infopanelUI}
\alias{mod_infopanelUI}
\title{mod_infopanelUI and mod_infopanel}
\usage{
mod_infopanelUI(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny module to generate the data tables
}
