% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapUI.R
\name{mod_mapUI}
\alias{mod_mapUI}
\alias{mod_map}
\title{mod_mapUI and mod_map}
\usage{
mod_mapUI(id)

mod_map(input, output, session, mod_data, mod_advancedFilters, ifndb)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{mod_data}{reactive with the reactive data and the data inputs}

\item{mod_advancedFilters}{reactive with the reactive values from the
advancedFilters module}

\item{ifndb}{pool with database connection object}
}
\description{
A shiny module to generate the data tables
}
