% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput_new.R
\name{mod_preFilInput}
\alias{mod_preFilInput}
\title{mod_preFilInput and mod_preFil}
\usage{
mod_preFilInput(id, nfidb)
}
\arguments{
\item{id}{shiny id}

\item{nfidb}{pool object to access the nfi database}
}
\description{
Module to create and populate the pre-filtering inputs
}
