% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableOutput.R
\name{mod_tableOutput}
\alias{mod_tableOutput}
\alias{mod_table}
\title{mod_tableOutput and mod_table}
\usage{
mod_tableOutput(id)

mod_table(input, output, session, mod_data, mod_advancedFilters, mod_map, ifndb)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{mod_data}{reactive with the reactive data and the data inputs}

\item{mod_advancedFilters}{reactive with the reactive values from
advancedFilters module}

\item{mod_map}{reactive with the mod_map inputs}

\item{ifndb}{db pool}
}
\description{
A shiny module to generate the base IFN plots table

mod_table server function
}
