% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vizInput.R
\name{mod_vizInput}
\alias{mod_vizInput}
\title{mod_vizInput and mod_viz}
\usage{
mod_vizInput(id, ifndb)
}
\arguments{
\item{id}{shiny id}

\item{ifndb}{pool object to access ifn db}
}
\description{
A shiny module to generate and populate the visualization inputs
}
