% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vizInput.R
\name{mod_viz}
\alias{mod_viz}
\title{mod_viz server function}
\usage{
mod_viz(input, output, session, mod_data, ifndb)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{mod_data}{reactive with the reactive data and the data inputs}

\item{ifndb}{pool object to access ifn db}
}
\description{
mod_viz server function
}
