% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{navbarPageWithInputs}
\alias{navbarPageWithInputs}
\title{nav bar with inputs}
\usage{
navbarPageWithInputs(..., inputs)
}
\arguments{
\item{...}{navbarpage arguments}

\item{inputs}{inputs objects}
}
\description{
from daattali in
https://github.com/daattali/advanced-shiny/tree/master/navbar-add-text
}
