% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{query_builder}
\alias{query_builder}
\title{query_builder}
\usage{
query_builder(mod_data, col_filter_expressions)
}
\arguments{
\item{mod_data}{reactives from data module}

\item{col_filter_expressions}{filter expressions for table module}
}
\description{
query_builder
}
