% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{table_data_modificator}
\alias{table_data_modificator}
\title{modifcate the table based on data inputs}
\usage{
table_data_modificator(data_scenario, scenario, admin_div, agg_level,
  diameter_classes, ifndb, updateProgress = NULL)
}
\arguments{
\item{data_scenario}{tbl_sql object}

\item{scenario}{character indicating the scenario}

\item{admin_div}{Input}

\item{agg_level}{Input}

\item{diameter_classes}{Input}

\item{ifndb}{pool object to access the ifn db}
}
\description{
this returns the data object to modify the table
}
