% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{infopanel_plot_gen}
\alias{infopanel_plot_gen}
\title{genreate the plot for the info panel}
\usage{
infopanel_plot_gen(data, click, color, viz_shape, agg_level,
  diameter_classes, tipo_grup_func, ifndb)
}
\arguments{
\item{data}{reactive data}

\item{click}{input}

\item{color}{input}

\item{viz_shape}{input}

\item{agg_level}{input}

\item{diameter_classes}{input}

\item{tipo_grup_func}{input}

\item{ifndb}{pool object to access the ifn db}
}
\description{
genreate the plot for the info panel
}
