% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{map_modificator}
\alias{map_modificator}
\title{modifcate the map based on vis inputs}
\usage{
map_modificator(data_scenario, scenario, ifn, inverse_pal, color, mida,
  tipo_grup_func, grup_func, statistic, admin_div, agg_level)
}
\arguments{
\item{statistic}{}
}
\description{
this returns the leaflet proxy object to modify the map
}
