% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_advancedFiltersUI.R
\name{mod_advancedFiltersUI}
\alias{mod_advancedFiltersUI}
\alias{mod_advancedFilters}
\title{mod_advancedFiltersUI and mod_advancedFilters}
\usage{
mod_advancedFiltersUI(id)

mod_advancedFilters(input, output, session, mod_data)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{mod_data}{mod_data reactives, to get when button to show advanced
filters is pressed}
}
\description{
A shiny module to process the advanced filters

mod_advancedFilters server function
}
