% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput_new.R
\name{mod_dataInput}
\alias{mod_dataInput}
\alias{mod_data}
\alias{mod_preFil}
\title{mod_dataInput and mod_data}
\usage{
mod_dataInput(id, nfidb)

mod_data(input, output, session, nfidb)

mod_preFil(input, output, session, nfidb)
}
\arguments{
\item{id}{shiny id}

\item{nfidb}{pool object to access the nfi database}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{nfidb}{}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{nfidb}{}
}
\description{
Module to create and populate the data inputs

mod_data server function

mod_preFil server function
}
