% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{mod_dataInput}
\alias{mod_dataInput}
\alias{mod_data}
\title{mod_dataInput and mod_data}
\usage{
mod_dataInput(id)

mod_data(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny module to create and populate the data inputs

mod_data server function
}
